changed the index to the new landing page
